CU-869625p3u Moved createShipment() method from Observer class to separate Shipment model
