+++ Observers/SalesOrderSaveBeforeObserver.php
@@ -5,13 +5,11 @@ use Magento\Sales\Api\Data\OrderInterface;
 use Magento\Sales\Api\OrderRepositoryInterface;
 use Magento\Framework\Event\Observer;
 use Magento\Framework\Event\ObserverInterface;
-use Magento\Framework\Message\ManagerInterface;
 use Magento\Framework\MessageQueue\PublisherInterface;
 use Magento\Framework\Serialize\Serializer\Json;
 use Magento\Sales\Model\Order;
 use StoreKeeper\StoreKeeper\Helper\Api\Auth;
 use StoreKeeper\StoreKeeper\Helper\Api\Orders as ApiOrders;
-use StoreKeeper\StoreKeeper\Logger\Logger;
 use StoreKeeper\StoreKeeper\Model\OrderSync\Shipment;
 
 class SalesOrderSaveBeforeObserver implements ObserverInterface
@@ -20,10 +18,7 @@ class SalesOrderSaveBeforeObserver implements ObserverInterface
     private Json $json;
     private PublisherInterface $publisher;
     private OrderRepositoryInterface $orderRepository;
-    private ApiOrders $apiOrders;
     private Shipment $shipment;
-    private Logger $logger;
-    private ManagerInterface $messageManager;
 
     /**
      * Constructor
@@ -32,29 +27,20 @@ class SalesOrderSaveBeforeObserver implements ObserverInterface
      * @param Json $json
      * @param PublisherInterface $publisher
      * @param OrderRepositoryInterface $orderRepository
-     * @param ApiOrders $apiOrders
      * @param Shipment $shipment
-     * @param Logger $logger
-     * @param ManagerInterface $messageManager
      */
     public function __construct(
         Auth $authHelper,
         Json $json,
         PublisherInterface $publisher,
         OrderRepositoryInterface $orderRepository,
-        ApiOrders $apiOrders,
-        Shipment $shipment,
-        Logger $logger,
-        ManagerInterface $messageManager
+        Shipment $shipment
     ) {
         $this->authHelper = $authHelper;
         $this->json = $json;
         $this->publisher = $publisher;
         $this->orderRepository = $orderRepository;
-        $this->apiOrders = $apiOrders;
         $this->shipment = $shipment;
-        $this->logger = $logger;
-        $this->messageManager = $messageManager;
     }
 
     /**
@@ -86,71 +72,10 @@ class SalesOrderSaveBeforeObserver implements ObserverInterface
                     );
 
                     if ($order->getStatus() == Order::STATE_COMPLETE) {
-                        $this->createShipment($order, $storeId);
+                        $this->shipment->createShipment($order, $storeId);
                     }
                 }
             }
         }
-    }
-
-    /**
-     * @param OrderInterface $order
-     * @param $storeId
-     * @return void
-     */
-    private function createShipment(OrderInterface $order, $storeId): void
-    {
-        $storekeeperId = $order->getStorekeeperId();
-
-        if ($this->apiOrders->allowShipmentCreation($order)) {
-            $shipmentData = [
-                'order_id' => $order->getId(),
-                'storekeeper_id' => $storekeeperId,
-                'store_id' => $storeId,
-                'items' => $this->getShipmentsItems($storeId, $storekeeperId)
-            ];
-
-            try {
-                $serializedData = $this->json->serialize($shipmentData);
-                $this->shipment->process($serializedData);
-                $order->setStorekeeperShipmentId($storekeeperId);
-            } catch (\Exception $e) {
-                $message = 'Error synchronizing shipment to StoreKeeper! Storekeeper order number: ' . $storekeeperId;
-                $this->logger->error(
-                    $message,
-                    [
-                        'error' =>  $this->logger->buildReportData($e),
-                        'request' =>  $serializedData
-                    ]
-                );
-
-                $this->messageManager->addNoticeMessage(__($message));
-            }
-        }
-    }
-    /**
-     * @param string $storeId
-     * @param string $storekeeperId
-     * @return array
-     */
-    private function getShipmentsItems(string $storeId, string $storekeeperId): array
-    {
-        $storeKeeperOrder = $this->apiOrders->getStoreKeeperOrder($storeId, $storekeeperId);
-        $shipmentItems = [];
-
-        if ($storeKeeperOrder) {
-            if (array_key_exists('order_items', $storeKeeperOrder)) {
-                foreach ($storeKeeperOrder['order_items'] as $orderItem) {
-                    if ($orderItem['is_shipping'] === false) {
-                        $shipmentItems[] = [
-                            'id' => $orderItem['id'],
-                            'quantity' => $orderItem['quantity']
-                        ];
-                    }
-                }
-            }
-        }
-
-        return $shipmentItems;
     }
 }
